Bugfix: bad pattern matching for alternative identifiers
DOIs can have the form "10.1051/0004-6361:20066170", which are matched
as identifier/section instead of directly as an alternative identifier

URLs such as:

https://ui.adsabs.harvard.edu/abs/10.1051/0004-6361:20066170

were returning a 404 (not found). Additionally, API queries did not wrap
identifiers in quotes, which lead to not finding such kind of DOIs.

Additionally, URLs such as
https://ui.adsabs.harvard.edu/abs/10.1051/0004-6361:20066170/abstract
were not working.

Finally, for any section, if an alternative identifier is used, it will
be redirected to the canonical bibcode.